IntegerField to BigIntegerField

+++ Registrations/models.py
@@ -7,7 +7,7 @@ class Register(models.Model):
     FirstName = models.CharField(max_length=100, blank=False)
     LastName = models.CharField(max_length=100, blank=False)
     email = models.EmailField(max_length=100, blank=False)
-    phone = models.IntegerField(blank=False)
+    phone = models.BigIntegerField(blank=False)
     password = models.CharField(max_length=200, blank=False)
     password2 = models.CharField(max_length=200, blank=False)
     Birthday_date = models.DateField(auto_now=False)
